requirements.txt
about-time==4.2.1
alive-progress==3.3.0
certifi==2022.9.24
charset-normalizer==2.1.1
deepdiff==6.2.1
dnspython==2.2.1
graphemeu==0.7.2
idna==3.4
mongoengine==0.24.2
ordered-set==4.1.0
pymongo==4.3.2
python-dotenv==0.21.0
requests==2.28.1
urllib3==1.26.12
